+++ run.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env sh
 mkdir -p build
 cd build
-cmake -DCMAKE_BUILD_TYPE=Debug -DENABLE_OPENSSL=ON -DBUILD_VERSION=0.0.5 ../
+cmake -DCMAKE_BUILD_TYPE=Debug -DENABLE_OPENSSL=ON -DBUILD_VERSION=0.0.6 ../
 make thinger_monitor
 if [ $THINGER_TOKEN ]; then
     THINGER_TOKEN="-t $THINGER_TOKEN"